% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfinitySparseMatrix.R
\name{as.InfinitySparseMatrix}
\alias{as.InfinitySparseMatrix}
\title{Convert an object to InfinitySparseMatrix}
\usage{
as.InfinitySparseMatrix(x)
}
\arguments{
\item{x}{An object which can be coerced into InfinitySparseMatrix, typically a matrix.}
}
\value{
An InfinitySparseMatrix
}
\description{
Convert an object to InfinitySparseMatrix
}
