% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfinitySparseMatrix.R
\name{as.list.BlockedInfinitySparseMatrix}
\alias{as.list.BlockedInfinitySparseMatrix}
\title{Splits a BlockedInfinitySparseMatrix into a list of
  InfinitySparseMatrices}
\usage{
\method{as.list}{BlockedInfinitySparseMatrix}(x, ...)
}
\arguments{
\item{x}{a BlockedInfinitySparseMatrix}

\item{...}{Ignored}
}
\value{
A list of InfinitySparseMatrices
}
\description{
Splits a BlockedInfinitySparseMatrix into a list of
  InfinitySparseMatrices
}
