% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptmatchS4.R
\name{as.optmatch}
\alias{as.optmatch}
\title{Function converting an S4-style Optmatch object into an equivalent s3 optmatch object, if possible}
\usage{
as.optmatch(Optmatch)
}
\description{
Function converting an S4-style Optmatch object into an equivalent s3 optmatch object, if possible
}
