% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfinitySparseMatrix.R
\name{subset.BlockedInfinitySparseMatrix}
\alias{subset.BlockedInfinitySparseMatrix}
\title{Subsetting for BlockedInfinitySparseMatrices}
\usage{
\method{subset}{BlockedInfinitySparseMatrix}(x, subset, select, ...)
}
\arguments{
\item{x}{BlockedInfinitySparseMatrix to be subset or bound.}

\item{subset}{Logical expression indicating rows to keep.}

\item{select}{Logical expression indicating columns to keep.}

\item{...}{Other arguments are ignored.}
}
\value{
If groups has names, a BlockedInfinitySparseMatrix with only
        the selected elements, otherwise an InfinitySparsematrix with
        only the selected elements
}
\description{
This matches the syntax and semantics of
subset for matrices.
}
