% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optmatchS3.R
\name{c.optmatch}
\alias{c.optmatch}
\title{Combine Optmatch objects}
\usage{
\method{c}{optmatch}(...)
}
\arguments{
\item{...}{Optmatch objects to be concatenated}
}
\value{
A combined Optmatch object
}
\description{
Combine Optmatch objects
}
