% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_reg_fm_prob.R
\name{cadlag_ceiling}
\alias{cadlag_ceiling}
\title{Right-continuous variant of base `ceiling()` function}
\usage{
cadlag_ceiling(x, tol = .Machine$double.neg.eps^0.5)
}
\arguments{
\item{x}{numeric}

\item{tol}{}
}
\value{
upwardly rounded version of x
}
\description{
Rounding function like ceiling -- except at the integers themselves,
where it's always one ahead.  (Right instead of left continuous;
thus "cadlag".)  Point being to map nonnegative doubles
to _positive_ integers, as required by fmatch.
}
\details{
The default tolerance value aims to steer well clear of inadvertently
passing 0L as where a 0 in x was being represented as a small negative
number, due to computer arithmetic.
}
\author{
Ben Hansen
}
\keyword{internal}
