% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasible.R
\name{caliperUpperBound}
\alias{caliperUpperBound}
\title{(Internal) Returns a reasonable upper bound on the arcs remaining after placing a caliper.}
\usage{
caliperUpperBound(scores, z, width, structure = NULL)
}
\arguments{
\item{scores}{A numeric vector of scores providing 1-D position of units}

\item{z}{Treatment indicator vector}

\item{width}{Width of caliper, must be positive.}

\item{structure}{Optional factor variable that groups the scores, as would
be used by \code{\link{exactMatch}}. Including structure allows for wider
calipers.}
}
\value{
numeric Total number of pairwise distances remaining after the caliper is placed.
}
\description{
(Internal) Returns a reasonable upper bound on the arcs remaining after placing a caliper.
}
\keyword{internal}
