% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{covariate_matched_dist_plot}
\alias{covariate_matched_dist_plot}
\title{Visualize the distances that remain after creating a stratified design
Visualize distances between units within sets after matching using a violin
plot, a boxplot, and dots. This function works best with covariates that
have more than two values. The sets with the largest and smallest absolute
differences are labeled by default.}
\usage{
covariate_matched_dist_plot(strata_indicator, trt, covariate)
}
\arguments{
\item{strata_indicator}{records membership in a strata. For example, it
could be the factor that is created by fullmatch or pairmatch from the
optmatch package. It should have names.}

\item{trt}{is a vector that indicates he treatment variable in a bipartite
matched design. It should be in the same order as the strata_indicator variable.}

\item{covariate}{is a vector with the same names as the strata_indicator and
node_color. The function will display distances within sets on this covariate.}
}
\value{
a ggplot object
}
\description{
Visualize the distances that remain after creating a stratified design
Visualize distances between units within sets after matching using a violin
plot, a boxplot, and dots. This function works best with covariates that
have more than two values. The sets with the largest and smallest absolute
differences are labeled by default.
}
\examples{
data(nuclearplants)
### Full matching on a Mahalanobis distance.
fm1 <- fullmatch(pr ~ t1 + t2, data = nuclearplants)

library(ggplot2)
library(tidyverse)

nuclearplants$fm1 <- fm1

## The below description shows the following about differences within set on t2:
## half of the sets differ by less than 7
## 90\% of the sets differ by less than 10.8
## The set with the largest difference had a difference of 13.
quantile(abs(t2_diffs$cov_diffs), c(0, .1, .25, .5, .9, 1))

## This next is a plot visualizing that distribution of within-stratum
## absolute differences

g0 <- covariate_matched_dist_plot(
  strata_indicator = fm1, trt = nuclearplants$pr,
  covariate = nuclearplants$t2
)
## Try the following to see the plot
## print(g0)
}
