% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfinitySparseMatrix.R
\docType{methods}
\name{dimnames,InfinitySparseMatrix-method}
\alias{dimnames,InfinitySparseMatrix-method}
\alias{dimnames<-,InfinitySparseMatrix,NULL-method}
\alias{dimnames<-,InfinitySparseMatrix,list-method}
\title{Get and set dimnames for InfinitySparseMatrix objects}
\usage{
\S4method{dimnames}{InfinitySparseMatrix}(x)

\S4method{dimnames}{InfinitySparseMatrix,list}(x) <- value

\S4method{dimnames}{InfinitySparseMatrix,`NULL`}(x) <- value
}
\arguments{
\item{x}{An InfinitySparseMatrix object.}

\item{value}{A list with two entries: the treated names and control names, respectively.}
}
\value{
A list with treated and control names.
}
\description{
InfinitySparseMatrix objects represent sparse matching problems
with treated units as rows of a matrix and controls units as
the columns of the matrix. The names of the units can be retrieved
and set using these methods.
}

