\name{DistanceSpecification-class}
\Rdversion{1.1}
\docType{class}
\alias{DistanceSpecification-class}

\title{Class \code{"DistanceSpecification"}}
\description{
  A \code{classUnion} for matrix like classes that can encode matching problems. Objects of this type are valid distance arguments for \code{\link{fullmatch}} and \code{\link{pairmatch}}. Users generally construct these objects through the \code{\link{match_on}} interface.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No public methods defined with class "DistanceSpecification" in the signature. Methods are generally defined for concrete classes, such as \code{matrix}.
}
\author{Mark M. Fredrickson}
\seealso{\code{\link{matrix}}, \code{\linkS4class{InfinitySparseMatrix}}, \code{\link{match_on}}}
\keyword{classes}
