% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.R
\name{edgelist}
\alias{edgelist}
\title{Create EdgeList object}
\usage{
edgelist(x, y = NULL)
}
\arguments{
\item{x}{object to convert to edgelist}

\item{y}{named character vector giving levels for nodes-representing columns, or NULL}
}
\value{
EdgeList
}
\description{
If y is a named character vector, then the names should
correspond to whatever in x would otherwise (i.e. if y were NULL)
translate to the levels set of the nodes-representing columns, while
the values themselves give the new levels.
}
\author{
Ben Hansen
}
\keyword{internal}
