% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementarySlackness.R
\name{evaluate_primal}
\alias{evaluate_primal}
\title{Compute value of primal problem given flows and arc costs}
\usage{
evaluate_primal(distances, solution)
}
\arguments{
\item{distances}{An InfinitySparseMatrix giving distances}

\item{solution}{A MCFSolutions object}
}
\value{
The value of the primal problem, i.e. sum of
products of \code{distances} with flow along arcs in \code{solution}
}
\description{
Compute value of primal problem given flows and arc costs
}
\author{
Hansen
}
