% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzDistanceSpecification.R
\name{findSubproblems}
\alias{findSubproblems}
\title{List subproblems of a distance}
\usage{
findSubproblems(d)
}
\arguments{
\item{d}{a distance specification}
}
\value{
list of distance specifications
}
\description{
Get all the subproblems from a distance specification
}
\author{
Mark M. Fredrickson
}
