% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasible.R
\name{fmla2treatedblocking}
\alias{fmla2treatedblocking}
\title{(Internal) A helper function to turn formulas into treatment and blocking variables}
\usage{
fmla2treatedblocking(x, ...)
}
\arguments{
\item{x}{A formula}

\item{...}{Arguments to be passed to model.frame (e.g. \code{data})}
}
\value{
data.frame containing two columns: \code{Z} is a treatment indicator,
\code{B} is a blocking factor
}
\description{
Given a function and any of the arguments normally passed to model.frame,
this function will return a data.frame with two columns: a treatment indicator
and a blocking factor.
}

