% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subproblemutils.R
\name{get_subproblem_info}
\alias{get_subproblem_info}
\title{Title}
\usage{
get_subproblem_info(object, type)
}
\arguments{
\item{object}{optmatch object}

\item{type}{either "subproblem" or "resolution"}
}
\description{
Title
}
