% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hash_dist}
\alias{hash_dist}
\alias{hash_dist.DenseMatrix}
\alias{hash_dist.matrix}
\alias{hash_dist.InfinitySparseMatrix}
\alias{hash_dist.BlockedInfinitySparseMatrix}
\alias{hash_dist.optmatch.dlist}
\title{(Internal) Hashing functions for various distance objects
#}
\usage{
hash_dist(dist)

\method{hash_dist}{DenseMatrix}(dist)

\method{hash_dist}{matrix}(dist)

\method{hash_dist}{InfinitySparseMatrix}(dist)

\method{hash_dist}{BlockedInfinitySparseMatrix}(dist)

\method{hash_dist}{optmatch.dlist}(dist)
}
\arguments{
\item{dist}{Distance object to hash. Must be one of
\code{InfinitySparseMatrix}, \code{BlockedInfinitySparseMatrix},
\code{DenseMatrix}, \code{matrix}, or \code{distmatch.dlist}.}
}
\value{
Hash on the distance object with a null \code{call}
}
\description{
(Internal) Hashing functions for various distance objects
#
}
\keyword{internal}
