% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfinitySparseMatrix.R
\docType{class}
\name{InfinitySparseMatrix-class}
\alias{InfinitySparseMatrix-class}
\title{Objects for sparse matching problems.}
\description{
\code{InfinitySparseMatrix} is a special class of distance specifications. Finite entries
indicate possible matches, while infinite entries indicated non-allowed
matches. This data type can be more space efficient for sparse matching
problems.
}
\details{
Usually, users will create distance specification using \code{\link{match_on}}, \code{\link{caliper}}, or
\code{\link{exactMatch}}.
}
\author{
Mark M. Fredrickson
}
\seealso{
\code{\link{match_on}}, \code{\link{caliper}}, \code{\link{exactMatch}}, \code{\link{fullmatch}},  \code{\link{pairmatch}}
}

