% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfinitySparseMatrix.R
\docType{methods}
\name{+,InfinitySparseMatrix,InfinitySparseMatrix-method}
\alias{+,InfinitySparseMatrix,InfinitySparseMatrix-method}
\alias{-,InfinitySparseMatrix,InfinitySparseMatrix-method}
\alias{*,InfinitySparseMatrix,InfinitySparseMatrix-method}
\alias{/,InfinitySparseMatrix,InfinitySparseMatrix-method}
\title{Element-wise addition}
\usage{
\S4method{+}{InfinitySparseMatrix,InfinitySparseMatrix}(e1, e2)

\S4method{-}{InfinitySparseMatrix,InfinitySparseMatrix}(e1, e2)

\S4method{*}{InfinitySparseMatrix,InfinitySparseMatrix}(e1, e2)

\S4method{/}{InfinitySparseMatrix,InfinitySparseMatrix}(e1, e2)
}
\arguments{
\item{e1}{an InfinitySparseMatrix object}

\item{e2}{an InfinitySparseMatrix object}
}
\value{
an InfinitySparseMatrix object representing
  the element-wise sum of the two ISM objects
}
\description{
\code{e1 + e2} returns the element-wise sum of
  two InfinitySparseMatrix objects.
  If either element is inf then
  the resulting element will be inf.

\code{e1 - e2} returns the element-wise subtraction of
  two InfinitySparseMatrix objects.
  If either element is inf then
  the resulting element will be inf.

\code{e1 * e2} returns the element-wise multiplication of
  two InfinitySparseMatrix objects.
  If either element is inf then
  the resulting element will be inf.

\code{e1 / e2} returns the element-wise division of
  two InfinitySparseMatrix objects.
  If either element is inf then
  the resulting element will be inf.
}
