% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEMON.R
\name{LEMON}
\alias{LEMON}
\title{(Internal) Helper function for accessing algorithms in LEMON solver}
\usage{
LEMON(algorithm = "CycleCancelling")
}
\arguments{
\item{algorithm}{LEMON algorithm to use. Choices are "CycleCancelling",
"CapacityScaling", "CostScaling", "NetworkSimplex". Default is
"CycleCancelling".}
}
\value{
String of the form "LEMON.<algorithm>"
}
\description{
(Internal) Helper function for accessing algorithms in LEMON solver
}
