% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptmatchS4.R
\name{makeOptmatch}
\alias{makeOptmatch}
\title{(Internal) Create \code{optmatch} objects, the result of matching.}
\usage{
makeOptmatch(distance, solutions, call, data = NULL)
}
\arguments{
\item{distance}{A \code{DistanceSpecificaton} object used to create the
match.}

\item{solutions}{A list of the results of the matching, one \code{list(cells,maxErr)} object per subproblem.}

\item{call}{The call to \code{fullmatch} or \code{pairmatch} to be displayed later.}

\item{data}{An object from which \code{names} or \code{row.names} will
provide the order of the items in the match. If no names are attached to this object, the contents will be used as names.}
}
\value{
\code{Optmatch} object with slots described in the class description.
}
\description{
This internal function is used to create the final output of the matching
functions (\code{\link{fullmatch}} and \code{\link{pairmatch}}). The
\code{Optmatch} object is an S4 object that inherits from a \code{factor}, but contains additional
information relating to the quality of the match in various slots.
}
\seealso{
\code{\link{summary.optmatch}}
}
\keyword{internal}
