% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_on.R
\name{match_on_szn_scale}
\alias{match_on_szn_scale}
\title{Pooled Dispersion for a Numeric Variable}
\usage{
match_on_szn_scale(x, trtgrp, standardizer = mad)
}
\arguments{
\item{x}{numeric variable}

\item{trtgrp}{logical or numeric.  If numeric, coerced to `T`/`F` via `!`}

\item{standardizer}{function or numeric of length 1}
}
\value{
numeric of length 1
}
\description{
Dispersion as pooled across a treatment and a control group. By default,
the measure of dispersion calculated within each group is not the
ordinary standard deviation but rather the robust alternative
provided by \code{stats::mad}.
}
\keyword{internal}
