% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCFSolutions.R
\name{c,SubProbInfo-method}
\alias{c,SubProbInfo-method}
\alias{c,NodeInfo-method}
\alias{c,ArcInfo-method}
\alias{c,MCFSolutions-method}
\alias{c,FullmatchMCFSolutions-method}
\title{Combine objects}
\usage{
\S4method{c}{SubProbInfo}(x, ...)

\S4method{c}{NodeInfo}(x, ...)

\S4method{c}{ArcInfo}(x, ...)

\S4method{c}{MCFSolutions}(x, ...)

\S4method{c}{FullmatchMCFSolutions}(x, ...)
}
\arguments{
\item{x}{object of particular class}

\item{...}{Various objects}
}
\value{
Combined objects
}
\description{
Combine objects
}
