% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.controls.cap.R, R/min.controls.cap.R
\name{maxControlsCap}
\alias{maxControlsCap}
\alias{minControlsCap}
\title{Set thinning and thickening caps for full matching}
\usage{
maxControlsCap(distance, min.controls = NULL, method = "RELAX-IV")

minControlsCap(distance, max.controls = NULL, method = "RELAX-IV")
}
\arguments{
\item{distance}{Either a matrix of non-negative, numeric
discrepancies, or a list of such matrices. (See
\code{\link{fullmatch}} for details.)}

\item{min.controls}{Optionally, set limits on the minimum number
of controls per matched set.  (Only makes sense for
\code{maxControlsCap}.)}

\item{method}{Choose which algorithm to use. Choices include "RELAX-IV" (default),
"CycleCancellingRunner", "CapacityScalingRunner", "CostScalingRunner",
"NetworkSimplexRunner". The last four are from the LEMON graph library. See this
site for details on their differences:
\url{https://lemon.cs.elte.hu/pub/doc/latest/a00606.html}.}

\item{max.controls}{Optionally, set limits on the maximum number
of controls per matched set.  (Only makes sense for
\code{minControlsCap}.)}
}
\value{
For \code{minControlsCap},
  \code{strictest.feasible.min.controls} and
  \code{given.max.controls}. For \code{maxControlsCap},
  \code{given.min.controls} and
  \code{strictest.feasible.max.controls}.

  \item{strictest.feasible.min.controls}{The largest values of the
  \code{\link{fullmatch}} argument \code{min.controls} that yield
  a full match;}
  \item{given.max.controls}{ The \code{max.controls} argument
  given to \code{minControlsCap} or, if none was given, a vector
  of \code{Inf}s. }
  \item{given.min.controls}{ The \code{min.controls} argument
  given to \code{maxControlsCap} or, if none was given, a vector
  of \code{0}s; }
  \item{strictest.feasible.max.controls}{The smallest values of
  the \code{\link{fullmatch}} argument \code{max.controls} that
  yield a full match.}
}
\description{
Functions to find the largest value of min.controls, or the
smallest value of max.controls, for which a full matching problem
is feasible.  These are determined by constraints embedded in the
matching problem's distance matrix.
}
\details{
The function works by repeated application of full matching, so on
large problems it can be time-consuming.
}
\note{
Essentially this is just a line search.  I've done several
  things to speed it up, but not everything that might be done.
  At present, not very thoroughly tested either: you might check
  the final results to make sure that \code{\link{fullmatch}}
  works with the values of \code{min.controls} (or
  \code{max.controls}) suggested by these functions, and that it
  ceases to work if you increase (decrease) those values.
  Comments appreciated.
}
\references{
Hansen, B.B. and S. Olsen Klopfer (2006),
  \sQuote{Optimal full matching and related designs via network
  flows}, \emph{Journal of Computational and Graphical Statistics}
  \bold{15}, 609--627.
}
\seealso{
\code{\link{fullmatch}}
}
\author{
Ben B. Hansen
}
\keyword{optimize}
