% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{missing_x_msg}
\alias{missing_x_msg}
\title{(Internal) If the x argument does not exist for
match_on, fullmatch, or pairmatch, use this function
to print a helpful message.}
\usage{
missing_x_msg(x_str, data_str, ...)
}
\arguments{
\item{x_str}{x as a string of code, usually deparse(substitute(x))}

\item{data_str}{data arg as string of code}

\item{...}{will look for 'z = <stuff>' in the extra args of caller}
}
\value{
string a helpful error message
}
\description{
(Internal) If the x argument does not exist for
match_on, fullmatch, or pairmatch, use this function
to print a helpful message.
}
\author{
Josh Buckner
}
