% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{optmatch-defunct}
\alias{optmatch-defunct}
\alias{pscore.dist}
\alias{mahal.dist}
\title{Functions deprecated or removed from optmatch}
\usage{
pscore.dist(...)

mahal.dist(...)
}
\arguments{
\item{...}{All arguments ignored.}
}
\description{
Over the course of time, several functions in optmatch have been removed in
favor of new interfaces and functions.
}
\seealso{
\code{\link{mdist}}, \code{\link{match_on}}
}
