% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptmatchS4.R
\name{optmatch_restrictions}
\alias{optmatch_restrictions}
\title{optmatch_restrictions}
\usage{
optmatch_restrictions(obj)
}
\arguments{
\item{obj}{An optmatch object}
}
\value{
A list of \code{min.controls}, \code{max.controls} and either
\code{omit.fraction} or \code{mean.controls}.
}
\description{
If \code{mean.controls} was explicitly specified in the creation of the
optmatch object, it is returned; otherwise \code{omit.fraction} is given.
}
\details{
Note that if the matching algorithm attempted to recover from initial
infeasible restrictions, the output from this function may not be the same as
the original function call.
}
