% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{predict.CBPS}
\alias{predict.CBPS}
\title{(Internal) Predict for CBPS objets}
\usage{
\method{predict}{CBPS}(x, newdata = NULL, ...)
}
\arguments{
\item{x}{A CBPS object}

\item{newdata}{Unused.}

\item{...}{Unused.}
}
\value{
Inverse logit of the fitted values.
}
\description{
(Internal) Predict for CBPS objets
}

