% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{predict_CBPS}
\alias{predict_CBPS}
\title{(Internal) Predict for CBPS objets}
\usage{
predict_CBPS(x)
}
\arguments{
\item{x}{A CBPS object}
}
\value{
Inverse logit of the fitted values.
}
\description{
(Internal) Predict for CBPS objets
}

