% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodecode.R
\name{prep_warm_nodes}
\alias{prep_warm_nodes}
\title{Function to handle preparation and extraction of node data. If information about nodes from a previously solved problem exists,
#' this function will extract that information and prepare it to be used in the following matching procedure. Currently is very much in development, not really ready for general use.}
\usage{
prep_warm_nodes(problems, old.node.data, old.prob.data)
}
\description{
Function to handle preparation and extraction of node data. If information about nodes from a previously solved problem exists,
#' this function will extract that information and prepare it to be used in the following matching procedure. Currently is very much in development, not really ready for general use.
}
