% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCFSolutions.R
\name{revise_ArcInfo_nodelabels}
\alias{revise_ArcInfo_nodelabels}
\title{Reset implicit node labels of an ArcInfo object}
\usage{
revise_ArcInfo_nodelabels(x, new, old_positions = 1L:length(new))
}
\arguments{
\item{x}{an ArcInfo object}

\item{new}{character; the new node labels (level sets for factors encoding arc start or end nodes)}

\item{old_positions}{integer; positions for old levels with new levels vector}
}
\value{
ArcInfo object with new levels
}
\description{
Reset implicit node labels of an ArcInfo object
}
\author{
Ben Hansen
}
\keyword{internal}
