% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasible.R
\name{setFeasibilityConstants}
\alias{setFeasibilityConstants}
\title{(Internal) Sets up the default values for maximum feasible problems}
\usage{
setFeasibilityConstants()
}
\description{
(Internal) Sets up the default values for maximum feasible problems
}
\keyword{internal}
