% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullmatch.R
\name{setTryRecovery}
\alias{setTryRecovery}
\title{(Internal) Sets up option to try recovery in \code{fullmatch}.}
\usage{
setTryRecovery()
}
\description{
(Internal) Sets up option to try recovery in \code{fullmatch}.
}

