% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfinitySparseMatrix.R
\docType{methods}
\name{show,BlockedInfinitySparseMatrix-method}
\alias{show,BlockedInfinitySparseMatrix-method}
\title{Displays a BlockedInfinitySparseMatrix}
\usage{
\S4method{show}{BlockedInfinitySparseMatrix}(object)
}
\arguments{
\item{object}{An BlockedInfinitySparseMatrix to print.}
}
\description{
Displays each block of the BISM separately.
}

