% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfinitySparseMatrix.R
\docType{methods}
\name{show,InfinitySparseMatrix-method}
\alias{show,InfinitySparseMatrix-method}
\title{Displays an InfinitySparseMatrix}
\usage{
\S4method{show}{InfinitySparseMatrix}(object)
}
\arguments{
\item{object}{An InfinitySparseMatrix to print.}
}
\description{
Specifically, displays an ISM by converting it to a matrix first.
}

