% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{stratification_graph}
\alias{stratification_graph}
\title{Visualize a matched design as graph}
\usage{
stratification_graph(strata_indicator, node_color)
}
\arguments{
\item{strata_indicator}{records membership in a strata. For example, it
could be the factor that is created by fullmatch or pairmatch from the
optmatch package. It should have names. The nodes will be labeled with these
names}

\item{node_color}{is a vector that will be used to color the nodes in the
graph. For example, it could be the treatment variable in a bipartite
matched design. It should be in the same order as the strata_indicator variable.}
}
\value{
a ggraph object
}
\description{
Visualize which units are connected to which other units after some
stratification using nodes and edges.
}
\examples{
data(nuclearplants)
### Full matching on a Mahalanobis distance.
fm1 <- fullmatch(pr ~ t1 + t2, data = nuclearplants)
g0 <- stratification_graph(fm1, node_color = nuclearplants$pr) + ggtitle("Full MH Stratification")
## Uncomment this next to see the graph
## print(g0)
}
