% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchfailed.R
\name{subproblemSuccess}
\alias{subproblemSuccess}
\title{(Internal) Report successful subproblems.}
\usage{
subproblemSuccess(x)
}
\arguments{
\item{x}{The result of \code{\link{fullmatch}} or \code{\link{pairmatch}}.}
}
\value{
A named logical vector indicating either success or failure for each subproblem.
}
\description{
\code{\link{fullmatch}} can break up a large matching problem into smaller
subproblems (for example, using strata defined by \code{\link{exactMatch}}).
This function lists the subproblems in a match and list whether at least on
treated unit was matched in subproblem. Subproblems that have no matched
treated units are said to have "failed."
}
\keyword{internal}
