% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optmatchS3.R
\name{update.optmatch}
\alias{update.optmatch}
\title{Performs an update on an \code{optmatch} object.}
\usage{
\method{update}{optmatch}(optmatch, ..., evaluate = TRUE)
}
\arguments{
\item{optmatch}{\code{Optmatch} object to update.}

\item{...}{Additional arguments to the call, or arguments with changed values.}

\item{evaluate}{If true evaluate the new call else return the call.}
}
\value{
An updated \code{optmatch} object.
}
\description{
NB: THIS CODE IS CURRENTLY VERY MUCH ALPHA AND SOMEWHAT UNTESTED, ESPECIALLY CALLING \code{update} ON AN
OPTMATCH OBJECT CREATED WITHOUT AN EXPLICIT DISTANCE ARGUMENT.
}
\details{
Note that passing \code{data} again is strongly recommended. A warning will be printed if the hash of the data used to generate the
\code{optmatch} object differs from the hash of the new \code{data}.
}
