% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfinitySparseMatrix.R
\name{cbind.InfinitySparseMatrix}
\alias{cbind.BlockedInfinitySparseMatrix}
\alias{cbind.InfinitySparseMatrix}
\alias{rbind.BlockedInfinitySparseMatrix}
\alias{rbind.InfinitySparseMatrix}
\title{Combine InfinitySparseMatrices or
  BlockedInfinitySparseMatrices by row or column}
\usage{
cbind.InfinitySparseMatrix(x, y, ...)

rbind.InfinitySparseMatrix(x, y, ...)

cbind.BlockedInfinitySparseMatrix(x, y, ...)

rbind.BlockedInfinitySparseMatrix(x, y, ...)
}
\arguments{
\item{x}{An InfinitySparseMatrix or BlockedInfinitySparseMatrix,
agreeing with \code{y} in the appropriate dimension.}

\item{y}{An InfinitySparseMatrix or BlockedInfinitySparseMatrix,
agreeing with \code{x} in the appropriate dimension.}

\item{...}{Other arguments ignored.}
}
\value{
A combined InfinitySparseMatrix or BlockedInfinitySparseMatrix
}
\description{
This matches the syntax and semantics of
cbind and rbind for matrices.
}
\author{
Mark Fredrickson
}

