% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dist_digest}
\alias{dist_digest}
\title{(Internal) Remove the call before digesting a distance so things
like omitting caliper and calling caliper=NULL give the same digest
#}
\usage{
dist_digest(dist)
}
\arguments{
\item{dist}{Distance object to hash. Must be one of
\code{InfinitySparseMatrix}, \code{BlockedInfinitySparseMatrix},
\code{DenseMatrix}, \code{matrix}, or \code{distmatch.dlist}.}
}
\value{
Hash on the distance object with a null \code{call}
}
\description{
(Internal) Remove the call before digesting a distance so things
like omitting caliper and calling caliper=NULL give the same digest
#
}
