% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmatch.R
\name{fmatch}
\alias{fmatch}
\title{Full matching via RELAX-IV min cost flow solver}
\usage{
fmatch(distance, max.row.units, max.col.units, min.col.units = 1,
  f = 1, node_prices = NULL)
}
\arguments{
\item{distance}{data frame w/ integer columns distance, treated, control}

\item{max.row.units}{numeric, upper limit on num treated units per matched set}

\item{max.col.units}{numeric, upper limit on num control units per matched set}

\item{min.col.units}{numeric, lower limit on num control units per matched set}

\item{f}{double, fraction of all row units to be matched}

\item{node_prices}{}
}
\description{
This is the function that calls the solver
}
\details{
Also handles pair matching, 1:k matching and matching
with between k>1 and l>k controls.  (If you want matching
with between k>1 and l>k members of the treatment group
per matched set, then you have to transpose the (virtual)
distance matrix before feeding the problem to this function.)
}
\author{
Ben Hansen, Mark Fredrickson, Adam Rauh
}
\keyword{internal}
