% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subproblemutils.R
\name{get_subproblem_info}
\alias{get_subproblem_info}
\title{Title}
\usage{
get_subproblem_info(object, type)
}
\arguments{
\item{object}{optmatch object}

\item{type}{either "subproblem" or any of the column names from the subproblems table associated with an MCFSolutions object (e.g. "resolution")}
}
\description{
Title
}
