% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subproblemutils.R
\name{get_subproblem_info}
\alias{get_subproblem_info}
\title{Get information about subproblems}
\usage{
get_subproblem_info(object, type)
}
\arguments{
\item{object}{optmatch object}

\item{type}{either "subproblem" or any of the column names from the subproblems table associated with an MCFSolutions object (e.g. "resolution")}
}
\value{
if \code{type = "subproblem"}, a named vector mapping units to subproblems is returned. If one or more columns from the subproblems table is specified, that column subset is returned. This might result in either a vector or data frame being returned.
}
\description{
Get information about subproblems
}
\examples{
data(nuclearplants)
f1 <- fullmatch(pr ~ cost + strata(pt), data=nuclearplants)
get_subproblem_info(object = f1, type = c("resolution", "groups"))
get_subproblem_info(object = f1, type = "subproblem")
}
