% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_on.R
\name{model.frame.svyglm}
\alias{model.frame.svyglm}
\title{This method quells a warning when \code{optmatch::scores()}
is applied to a svyglm object.  I don't expect that it to be
useful in other contexts, only exporting it for ease of debugging.}
\usage{
\method{model.frame}{svyglm}(formula, ...)
}
\description{
This method quells a warning when \code{optmatch::scores()}
is applied to a svyglm object.  I don't expect that it to be
useful in other contexts, only exporting it for ease of debugging.
}
\keyword{internal}
