% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{optmatch-defunct}
\alias{mahal.dist}
\alias{optmatch-defunct}
\alias{pscore.dist}
\title{Functions deprecated or removed from optmatch}
\usage{
pscore.dist(...)

mahal.dist(...)
}
\arguments{
\item{...}{All arguments ignored.}
}
\description{
Over the course of time, several functions in optmatch have been removed in
favor of new interfaces and functions.

All functionality of the \code{pscore.dist} function has been moved into to
the \code{\link{mdist}} function. Additionally, this function will also act on
other objects, such as formulas.
The \code{\link{match_on}} function
also provides similar functionality, though with a different syntax.

All functionality of the \code{mahal.dist} function has been moved into to
the \code{\link{mdist}} function. Additionally, this function will also act on
other objects, such as \code{glm} objects. The \code{\link{match_on}} function
also provides similar functionality, though with a different syntax.
}
\seealso{
\code{\link{mdist}}, \code{\link{match_on}}
}

