% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plantdist}
\alias{plantdist}
\title{Dissimilarities of Some U.S. Nuclear Plants}
\format{A matrix with 7 rows and 19 columns}
\source{
The data appear in Cox, D.R. and Snell, E.J. (1981),
  \emph{Applied Statistics: Principles and Examples}, p.82 (Chapman
  and Hall), and are due to W.E. Mooz.
}
\usage{
plantdist
}
\description{
This matrix gives discrepancies between light water nuclear power
plants of two types, seven built on the site of an existing plant
and 19 built on new sites.  The discrepancies summarize differences
in two covariates that are predictive of the cost of building a
plant.
}
\references{
Rosenbaum, P.R. (2002), \emph{Observational Studies},
  Second Edition, p.307 (Springer).
}
\keyword{dataset}

