% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{predict.CBPS}
\alias{predict.CBPS}
\title{(Internal) Predict for CBPS objects}
\usage{
\method{predict}{CBPS}(x, newdata = NULL, type = c("link", "response"), ...)
}
\arguments{
\item{x}{A CBPS object}

\item{newdata}{Unused.}

\item{type}{Return inverse logits of fitted values (the default) or fitted values themselves}

\item{...}{Unused.}
}
\value{
Inverse logit of the fitted values.
}
\description{
The CBPS package fits \sQuote{covariate balancing propensity score} for use in propensity score
weighting.  In the binary treatment case they can also be used for matching.  This method helps to 
implement that process in a manner consistent with use of propensity scores elsewhere in optmatch; see
\code{\link{scores}} documentation.
}

