\name{relaxinfo}
\alias{relaxinfo}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Display license information about embedded code }
\description{
Function to display license information regarding code embedded
  in \code{optmatch}
}
\usage{
relaxinfo()
}
%- maybe also 'usage' for other objects documented here.
%\details{
%}
\value{
Some information about licenses of code and algorithms on which
\code{fullmatch} depends.
}
\author{Ben B. Hansen }

\keyword{optimize}% at least one, from doc/KEYWORDS
