% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_on.R
\name{scoreCaliper}
\alias{scoreCaliper}
\title{(Internal) Helper function to create an InfinitySparseMatrix from a set of
scores, a treatment indicator, and a caliper width.}
\usage{
scoreCaliper(x, z, caliper)
}
\arguments{
\item{x}{The scores, a vector indicating the 1-D location of each
unit.}

\item{z}{The treatment assignment vector (same length as \code{x})}

\item{caliper}{The width of the caliper with respect to the scores
\code{x}.}
}
\value{
An \code{InfinitySparseMatrix} object, suitable to be
  passed to \code{\link{match_on}} as an \code{within} argument.
}
\description{
(Internal) Helper function to create an InfinitySparseMatrix from a set of
scores, a treatment indicator, and a caliper width.
}

